chore: move isFipsRegion to regionConfig
