+++ packages/config-resolver/src/regionConfig/isFipsRegion.spec.ts
